Rename SetPropertyFactory.CodeGenerator -> SetProperty
SetPropertyFactory is now SetProperty.Factory
